Move the process reaper into its own file
